+++ amazon/evidently/client/src/main/kotlin/org/http4k/connect/amazon/evidently/actions/LegacyReferences.kt
@@ -1,0 +1,4 @@
+package org.http4k.connect.amazon.evidently.actions
+
+// TODO Remove when breaking changes acceptable
+typealias VariableValue = org.http4k.connect.amazon.evidently.model.VariableValue